efivar: Split types and guids out into a different header.
