+++ src/libefivar.map.in
@@ -136,6 +136,7 @@ LIBEFIVAR_1.38 {
 		efi_set_loglevel;
 		efi_get_libefivar_version;
 		efi_guid_grub;
+		efi_guid_external_management;
 		efi_variable_alloc;
 		efi_variable_export_dmpstore;
 } LIBEFIVAR_1.37;
